backend/pyproject.toml
[tool.poetry]
name = "kokos"
version = "0.1.0"
description = ""
authors = ["4got10king <Aldavletshina77654@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
urllib3 = "^2.2.3"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.5.2"
fastapi-cache2 = "^0.2.2"
redis = "^5.1.1"
python-jose = "^3.3.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.21" }
jinja2 = "^3.1.2"
sqladmin = "0.16.1"
asyncpg = "^0.29.0"
wtforms = "^3.1.2"
itsdangerous = "^2.2.0"
alembic = "^1.13.3"
alembic-postgresql-enum = "^1.3.0"
pyjwt = "^2.9.0"
cyrtranslit = "^1.1.1"
transliterate = "^1.10.2"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
bcrypt = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
